Move story detail cell creation into its own controller
